refactor: convert test files to use path aliases
Convert all test file imports from relative paths to path aliases for consistency and maintainability.

Changes:
- Updated 13 test files to use @/ path aliases
- Replaced relative imports (./component, ../../utils) with absolute path aliases
- Improved code consistency with project architecture standards

Files modified:
- src/actions/mailchimp-{reports,root}.test.ts
- src/components/dashboard/reports*.test.tsx
- src/components/mailchimp/**/*.test.tsx
- src/components/ui/*-badge.test.tsx
- src/utils/**/*.test.ts

Benefits:
- Easier refactoring when moving test files
- Consistent with project's path alias enforcement
- Cleaner, more maintainable import statements

All tests pass (360/360 ✓)
