+++ src/components/mailchimp/lists/list-card.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@/test/test-utils";
 import { expectNoA11yViolations, renderWithA11y } from "@/test/axe-helper";
-import { ListCard } from "./list-card";
+import { ListCard } from "@/components/mailchimp/lists";
 import type { List } from "@/services";
 
 const mockList: List = {
