+++ src/components/mailchimp/lists/list-stats.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@/test/test-utils";
 import { expectNoA11yViolations, renderWithA11y } from "@/test/axe-helper";
-import { ListStats } from "./list-stats";
+import { ListStats } from "@/components/mailchimp/lists";
 import type { DashboardListStats as ListStatsType } from "@/types/mailchimp/lists";
 
 const mockStats: ListStatsType = {
